+++ commands/systems/update_account.py
@@ -82,7 +82,7 @@ class CommandHandler(CommandHandlerBase):
             return
 
         fullurl = url + username
-        link = UrlAccess.process_request(redfishConfig, UrlStatus(fullurl), 'PATCH', True, json.dumps(JsonBuilder.getElement('main'), indent=4))
+        link = UrlAccess.process_request(redfishConfig, UrlStatus(fullurl), 'PATCH', True, JsonBuilder.getElement('main'))
 
         Trace.log(TraceLevel.INFO, '   -- {0: <14}: {1}'.format('Status', link.urlStatus))
         Trace.log(TraceLevel.INFO, '   -- {0: <14}: {1}'.format('Reason', link.urlReason))