+++ core/trace.py
@@ -57,12 +57,14 @@ class Trace:
                 cls.tracelevel = newlevel
         except Exception as e:
             Trace.log(TraceLevel.ERROR, '   -- Unable to set trace level ({}) for cls {}: Exception: {}'.format(newlevel, cls, str(e)))
-            pass
 
     @classmethod
     def getlevel(cls):
         return cls.tracelevel, cls.tracelabels[cls.tracelevel]
 
+    @classmethod
+    def getlevelint(cls):
+        return cls.tracelevel
 
     @classmethod
     def log(cls, level, entry):